Add some inline comments
